모든 Id String으로 받지 말고 long으로 받음

+++ data/src/main/java/com/inkiu/data/mapper/MediaDataToEntityMapper.kt
@@ -7,7 +7,7 @@ class MediaDataToEntityMapper : Mapper<Media, MediaEntity> {
 
     override fun map(src: Media): MediaEntity {
         return MediaEntity(
-            id = src.id.toLong(),
+            id = src.id,
             mediaUrl = src.url,
             type = getType(src.type)
         )
